// SMC.Academico.Client/projects/smc-sga-administrativo/src/app/areas/apr/situacao-aula-lote/models/situacao-aula-lote-lista.model.ts
export interface SituacaoAulaLoteListaModel {
  seq: Number;
  descricaoDivisaoTurma: string;
  descricaoColaboradores: string;
  dataHorario: string;
  qtdAlunoFalta: string;
  prazoApuracao: string;
  descricaoNivelEnsino: string;
  prazoAlteracao: string;
  codigoDivisaoTurma: string;
  alunosHistoricoEscolar: boolean;
}